requirements.txt
asgiref==3.5.2
Django==4.0.5
sqlparse==0.4.2
psycopg2-binary==2.9.3
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
flake8 >= 3.0.0
unittest-xml-reporting==3.2.0
django-environ==0.4.0
Pillow==9.1.1
djangorestframework==3.13.1
django-oauth-toolkit==2.1.0
drf-spectacular==0.22.1